cast flag to int

+++ plant_doc/utils/leaf_image.py
@@ -11,7 +11,7 @@ def prepare(rgb_img, gray_low, gray_high):
     logger.debug('received rbg image')
     gray_img = pcv.rgb2gray(rgb_img=rgb_img)
     logger.debug('converted rbg image to grayscale image')
-    masked_img = cv2.inRange(gray_img, np.float32(gray_low), np.float32(gray_high))
+    masked_img = cv2.inRange(gray_img, gray_low, gray_high)
     logger.debug('filtered grayscale image')
     fill_image = pcv.fill_holes(masked_img)
     logger.debug('filled holes in binary mask')
